+++ lib/features/record/components/announcement_bar/announcement_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:pilll/features/record/components/announcement_bar/components/admob.dart';
+import 'package:pilll/features/record/components/announcement_bar/components/admob_native_advanced.dart';
 import 'package:pilll/utils/analytics.dart';
 import 'package:pilll/provider/pill_sheet_group.dart';
 import 'package:pilll/provider/pilll_ads.dart';
@@ -122,7 +122,16 @@ class AnnouncementBar extends HookConsumerWidget {
         return PilllAdsAnnouncementBar(pilllAds: pilllAds, onClose: () => showPremiumIntroductionSheet(context));
       }
 
-      return const AdMobNativeAdvance();
+      // iPhone mini, iPhone SEサイズだとNativeAdvanceの広告の高さだと画面占領が目立つので高さで広告を分岐する
+      // SEもminiも将来的にサポート対象外(どちらも生産終了)しているので、この分岐は将来的には不要になる
+      // iPhone SEはiOS15までサポートされているので、2024-01以降にiOSのバージョンを16に上げて回避するでも良さそう
+      // ただ、一時的にサポートするコードを書いてしまっているので、iOSのバージョンを上げるのは必須じゃない。2024-01以降で他に困ることがあればiOSのバージョンを上げる
+      // mini: 780, SE: 667
+      if (MediaQuery.of(context).size.height <= 780) {
+        return const AdMobNativeAdvance();
+      } else {
+        return const AdMobNativeAdvance();
+      }
     } else {
       final shownRecommendSignupNotificationForPremium = () {
         if (isLinkedLoginProvider) {
